Kill command checking
When checking the kill command, if you check that the command is "!kill" AND if the userID is zacimac's then perform the subsequent code. Otherwise, if one of the checks fails, the statement will drop into the else block, thus letting the user know they cannot use this command. Personally, you should let it silently fail by taking out the "else{...}" altogether.